Implement EventBus
The EventBus class enables other actors to subscribe to specific events and receive
notifications when these events are published to the event bus. Here is an example of
how to use this functionality:

```ruby
  event_bus = system.actor_of(EventBus.props)
  event_bus.tell EventBus::Subscribe.new(event: Numeric, subscriber: actor_ref2)
  event_bus.tell EventBus::Subscribe.new(event: Float, subscriber: actor_ref2)
  event_bus.tell EventBus::Publish.new(event: 42.2)
```

In the above example, only +actor_ref2+ receives the +42.2+ message. However, if you publish +424+
it will be received by both actors.